+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.3.0
+
+### Minor Changes
+
+-   c17bc130: When passing a default UI5 version its availability will be verified before offering as a choice
+
+### Patch Changes
+
+-   Updated dependencies [c17bc130]
+    -   @sap-ux/inquirer-common@0.2.3
+
+## 0.2.8
+
+### Patch Changes
+
+-   61b46bc8: Security upgrade fixes
+-   Updated dependencies [61b46bc8]
+    -   @sap-ux/ui5-info@0.3.3
+    -   @sap-ux/inquirer-common@0.2.2
+
 ## 0.2.7
 
 ### Patch Changes